Add cgroup v2 support
cgroup v1 uses two special files to determine access, where cgroup v2
uses eBPF programs to control access. The code will attach a custom eBPF
program which allows run-time reconfiguration and detach docker's default.

eBPF programs will be detached when the attaching program dies, which
can be dangerous if container-hotplug exits unexpectedly while the program
is running, so we instead pin it (so it stays when the program exits) and
unpin it after the docker container is down. In this case we might have
garbage eBPF programs pinned when container-hotplug exits unexpectedly but
it is safe.