Use upstream udev-rs
The fork provides `Send`/`Sync` impl to udev types, but according to
libudev docs, an udev instance created on a thread must not be used or
freed in another thread, even with locking. So switch to upstream udev-rs
which does not have `Send`/`Sync` impl.

To accommodate this change, `LocalBoxStream` is used instead of `BoxStream`
since the latter requires `Send`.