Fix slow process exit issue
The slow process exit is caused by two issues.

1. Blocking tokio operations.

This can either be
caused by `spawn_blocking` (which is internally used by a lot of tokio
std operations) and using block_on.

The "guard" pattern can especially causes this because tokio is not
expecting to block in a runtime thread.

2. Not receiving destroy event.

We listen for the destroy event to determine if the removal is
successful. However we only `.await` it after sending the removal
call. So if the removal takes a short time, then the container
could be destroyed before we start listening (since future does
nothing until polled). This is fixed by spawning.