Refactor root device unplug handling
Currently, when the root device is unplugged, we send a TERM signal and
immediately ask docker to force remove the device. This is effectively
sending a KILL signal since in almost all cases the container does not have
enough time to perform graceful cleanup between TERM and force removal,
which will send a KILL signal. So we could instead just send a KILL signal
directly and wait for exit.

With this change, now for all `Detach` and `Stopped` scenario, we know
the container is stopped, and since we passed in `--rm` to docker, we know
that docker will perform container removal itself. This means that we do
not need to explicitly invoke remove anymore and this allows massive cleanup.